requirements.txt
django >3.2,<3.3
python-decouple
pillow

djangorestframework
markdown       
django-filter